Election analysis done through VS code
Using VS code, I found the following: total amount and percentage of votes in each counties and the city with the largest voter turnout rate. In the same file, I looked at each of the candidates, and the total amount and percentage of votes they received. Lastly, the winning candidate was revealed with their vote count and percentage.